all done, but still have to split the function

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: aboudoun <aboudoun@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/12/13 15:20:12 by aboudoun          #+#    #+#             */
-/*   Updated: 2021/12/17 18:44:30 by aboudoun         ###   ########.fr       */
+/*   Updated: 2021/12/18 14:19:36 by aboudoun         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ int		ft_printf(const char *s, ... );
 void	ft_putchar(int c, int *len);
 void	ft_putstr(char *str, int *len);
 void	ft_putnbr(long nb , int *len);
-void    ft_puthex(long nbr, int *len, char flag);
-void ft_putptr(unsigned long long ptr, int *len);
+void    ft_puthex(unsigned long nbr, int *len, char flag);
+void 	ft_putptr(unsigned long ptr, int *len);
 
 #endif
